Variable depth of conversion

+++ Geant_simulation/include/Hits.h
@@ -9,7 +9,8 @@ public:
 	Hits(int);
 	~Hits();
 
-	vector<int> N_reg_v;
+	//vector<int> N_reg_v;
+	vector<double> N_reg_v;
 	vector<double> xpos_v;
 	vector<double> ypos_v;
 };
